Moved question types and common shared components
