+++ packages/web-forms/src/components/common/CheckboxWidget.vue
@@ -1,5 +1,5 @@
 <script lang="ts" setup>
-import TextMedia from '@/components/widgets/TextMedia.vue';
+import TextMedia from '@/components/common/TextMedia.vue';
 import { selectOptionId } from '@/lib/format/selectOptionId.ts';
 import type { SelectNode } from '@getodk/xforms-engine';
 import Checkbox from 'primevue/checkbox';
