+++ packages/web-forms/src/components/common/IconSVG.vue
