+++ packages/web-forms/src/components/common/ImageBlock.vue
