+++ packages/web-forms/src/components/common/MultiselectDropdown.vue
@@ -61,7 +61,7 @@ if (props.question.appearances['no-buttons']) {
 </template>
 
 <style scoped lang="scss">
-@use 'primeflex/core/_variables.scss' as pf;
+@use '../../../../../node_modules/primeflex/core/variables' as pf;
 
 .multi-select-dropdown {
 	width: 100%;
