+++ packages/web-forms/src/components/question-types/input/InputControl.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import InputDate from '@/components/controls/Input/InputDate.vue';
+import InputDate from '@/components/question-types/Input/InputDate.vue';
 import type { AnyInputNode } from '@getodk/xforms-engine';
 import { computed, inject, provide, ref } from 'vue';
 import ControlText from '../../ControlText.vue';
