+++ packages/web-forms/src/components/question-types/input/InputNumeric.vue
