+++ packages/web-forms/src/components/question-types/range/RangeControl.vue
