+++ packages/web-forms/src/components/question-types/range/RangeSlider.vue
