+++ packages/web-forms/src/components/question-types/upload/UploadImagePreview.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
-import IconSVG from '@/components/widgets/IconSVG.vue';
-import ImageBlock from '@/components/widgets/ImageBlock.vue';
+import IconSVG from '@/components/common/IconSVG.vue';
+import ImageBlock from '@/components/common/ImageBlock.vue';
 import type { ObjectURL } from '@getodk/common/lib/web-compat/url.ts';
 import { createObjectURL, revokeObjectURL } from '@getodk/common/lib/web-compat/url.ts';
 import type { UploadNode } from '@getodk/xforms-engine';
